Fix DriverExistence command
Due to a design issue, DriverExistence command, the command that allows
userspace to check the existence of the pattgen capsule on the platform,
was not recognized correctly. The issue came from the fact that
DriverExistence was categorized as PattgenCommand, commands that act on
pattern generator channels. This is clearly wrong and fixed by this
commit.

Signed-off-by: Ioan-Cristian CÎRSTEA <ioan.cirstea@oxidos.io>