Handle the case when a process detaining the lock no longer exists.
Previously, the capsule did not handle the case when a process detaining
the lock terminates or faults before releasing the lock. This is a
problem because the lock will never get released, blocking other
processes waiting for the lock. Now, the capsule checks if the process
is alive or not. If not, the previous owner of the lock is ignored until
the next successful lock command.