ResetManager capsule and driver (#47)
* ResetManager capsule and driver

cleaned up warnings from make prepush

* Debug build needs LTO enabled and vfunc elimination disabled

* capsule command for getting reset dump info

* resolved issues and typos raised in RyanTorok's review

* Started working on OTP.

Added constructor + init method.

* First iteration of Pattgen

It is not tested yet.

* Finished OTP initialization.

Check timeout doesn't work as expected. The kernel doesn't print any
message anymore, so I suspect the main loop is not reached anymore and
check timeout blocks the chip.

* Implemented read operation

* Added tests

* Added support for 64-bit reads

* Added pattgen interrupts + tests

* Fixed timeout issue + test only buferred partitions

* Added docs

* Small improvements

* OTP uses field-based read

* Fixed endianess problem

* Added documentation

* Revert endianess problem + added test for read_en_sram_ifetch()

* Replaced helper macros with functions

* Change the signature of set_check_timeout() method.

As suggested by Ryan Torok, I changed the signature of the
set_check_timeout() method. Instead of passing `timeout` as u32, it is
now passed as `Option<NonZeroU32>`. `Some` variant means that timeout is
enabled with the inner value, while `None` means that timeout should be
disabled.

* Run make prepush

Run make prepush to prepare the code for push.

* Add support for conditional compilation for tests submodule

Now, to use any code from the ̀`tests` submodule, the corresponding
feature must be enabled. This helps avoid including any test code into
production code.

* Add abstraction types

This commit adds abstraction types that will be used to develop the
flash peripherals:

+ Bank: flash bank
+ FifoLevel: level of flash FIFO
+ FlashAddress: address in flash address space
+ InfoPartitionType: type of an info partition type
+ PageIndex: index of a page within a bank
+ PagePosition: position of a page represented by bank and index
+ Page: a flash page

It also adds a minimal test framework. Since no code deals with hardware
or Tock, tests can be run directly on the host machine via `cargo test`.

Signed-off-by: Ioan-Cristian CÎRSTEA <ioan.cirstea@oxidos.io>

* Add conditional compilation for tests submodule

To include tests submodule, it is now required to enable the `tests`
feature from `boards` crate. This allows production code to avoid using
any test code.

* Fix compilation error + improve test module documentation.

Due to naming error, the code didn't compile. This commit fixes this.

Also, this commit improves the documentation for the pattgen test
module, explicitly stating that the "tests" feature must be enabled for
the lowrisc dependency in order to test the pattgen.

* Started reworking the USB driver

This commit adds several abstraction layers for the USB driver. However,
the driver panics at first interrupt, since interrupts are not yet
handled.

Signed-off-by: Ioan-Cristian CÎRSTEA <ioan.cirstea@oxidos.io>

* First implementation of the pattgen capsule

All basic functionality implemented apart from pattern generation
completion upcalls.

Signed-off-by: Ioan-Cristian CÎRSTEA <ioan.cirstea@oxidos.io>

* Add support for pattern generation done upcall

The pattgen capsule now supports pattern generation done upcalls.
Pattern generation done upcalls are issued to processes when pattern
generation finishes. The process is notified of the pattgen channel that
has just completed pattern generation.

Signed-off-by: Ioan-Cristian CÎRSTEA <ioan.cirstea@oxidos.io>

* Fix DriverExistence command

Due to a design issue, DriverExistence command, the command that allows
userspace to check the existence of the pattgen capsule on the platform,
was not recognized correctly. The issue came from the fact that
DriverExistence was categorized as PattgenCommand, commands that act on
pattern generator channels. This is clearly wrong and fixed by this
commit.

Signed-off-by: Ioan-Cristian CÎRSTEA <ioan.cirstea@oxidos.io>

* Handle the case when a process detaining the lock no longer exists.

Previously, the capsule did not handle the case when a process detaining
the lock terminates or faults before releasing the lock. This is a
problem because the lock will never get released, blocking other
processes waiting for the lock. Now, the capsule checks if the process
is alive or not. If not, the previous owner of the lock is ignored until
the next successful lock command.

* Added support for short get control commands.

This commit adds support for short get control commands, i.e. SETUP
packets followed by IN packets. Currently, if the packet doesn't fit in
64B, it is truncated. The driver panics `set_address()` method.

Signed-off-by: Ioan-Cristian CÎRSTEA <ioan.cirstea@oxidos.io>

* Handle large data control IN transactions.

This commits adds support for large data control IN transactions
composed of several USB packets.

Signed-off-by: Ioan-Cristian CÎRSTEA <ioan.cirstea@oxidos.io>

* Add support for unlock upcall

Before this commit, processes could only busy-wait when attempting to
acquire pattgen lock. Now, they have the option to register an upcall to
be notified when the lock is released.

Signed-off-by: Ioan-Cristian CÎRSTEA <ioan.cirstea@oxidos.io>

* Added support for set address requests

The driver is now able to handle set address requests. Now, it panics on
other set requests, such as set configuration.

Signed-off-by: Ioan-Cristian CÎRSTEA <ioan.cirstea@oxidos.io>

* Cleaned issues pointed out by jon-flatley

* ResetManager capsule and driver

Rebased

* Debug build needs LTO enabled and vfunc elimination disabled

* capsule command for getting reset dump info

* resolved issues and typos raised in RyanTorok's review

* Cleaned issues pointed out by jon-flatley

* corrected conflict in earlgrey/chips.rs

---------

Signed-off-by: Ioan-Cristian CÎRSTEA <ioan.cirstea@oxidos.io>
Co-authored-by: Ioan-Cristian CÎRSTEA <ioan.cirstea@oxidos.io>
Co-authored-by: Ioan-Cristian CÎRSTEA <ioan.cirstea@stud.acs.upb.ro>
Co-authored-by: Jon Flatley <47988881+jon-flatley@users.noreply.github.com>
Co-authored-by: Ioan-Cristian <118638518+Ioan-Cristian@users.noreply.github.com>