Fix flash controller initialization
When using ROM_EXT, the flash operation done interrupt is set. This
caused a spurious trigger of the flash interrupt handler which
ultimately leds to kernel panic.

The error code register and status are also set. They need to be
cleared, as the flash controller driver assumes register reset values.