[otbn] add redirect of LLVM __udivdi3 function
OpenTitan cryptolib depends on opentitan:sw/device/lib/base:math, which
black-holes this function in favor of udiv64_slow for code size reasons.
The redirect from Tock is necessary to prevent the linker hitting the
black hole when linking to cryptolib.

This commit causes any division operations in the OpenTitan boards not
optimized out by the compiler or with a power-of-two divisor to call
udiv64_slow.