[spi_host] fix interrupt loop bug in spi_host
The spi_host driver erroneously enabled the TX_EMPTY status event on
initialization, causing an interrupt loop, whereas it should only be
enabled when we are actually trying to send something. This issue was
covered up on earlgrey_es_sival because the init() code was writing a
single byte to work around an issue with the TXQD counter register.