[opentitan] Support oneshot digests with cryptolib
Adds and implements a new set of HILs for performing oneshot hash
operations directly on inputs from the Grant region. This is necessary
for supporting the OpenTitan KMAC block, which does not support
streaming usage. It also yields performance gains for operations with
the HMAC block, as cryptolib does not need to copy context information
in and out of the HMAC block between the update() and finalize()
operations when using the oneshot API.

Due to kernel limitations on dynamic memory allocation, userspace
applications are required to adhere to certain sizing and alignment
requirements for allowed buffers for the digest when using operations
with a dynamically-sized output (KMAC, SHAKE, cSHAKE). Cryptolib
requires 4-byte alignment for all digest buffers, and allowing an
unaligned buffer or one whose length is not a multiple of 4 bytes will
cause an error to be returned.