[opentitan] remove hard-coded register definitions
Removes the usage of hard-coded register definitions in OpenTitan
drivers. HMAC, OTBN, and CSRNG were not corrected because they are
expected to be phased out in favor of cryptolib-based drivers, and the
fixes would not have been trivial.

In the OpenTitan tree, builds now use auto-generated register defintions
that are synced with the RTL hjson files. In the Tock tree, the drivers
still use vendored copies of these files. The latter will eventually be
phased out in favor of building everything from the OpenTitan tree.