[pwrmgr] Temporarily disable sleeping
\#197 fixed an misalignment bug in the power manager register map that
was causing the driver implementation to think it was enabling chip
sleeping, but was actually writing to an irrelevant register instead.
Once sleeping was properly enabled, it was discovered that a proper
configuration of when to wake up was never implemented because the above
bug was covering up any issues.

Until a wakeup configuration is designed, this PR temporarily disables
sleeping to prevent the chip hanging due to lack of wakeup.