[configurator] Optionally enable peripherals according to configuration
Adds a tracer to the configurator's chips/lowrisc crate, which allows the
configurator to determine at the board generation step if each driver should be
included in the board.

When a meta-driver is included as a dependency, the
meta-driver's 'trace_dependencies' function has the choice to:

1) Specify that one or more peripherals need their peripheral driver included.
2) Specify that one or more peripherals need interrupts enabled.

For example, oneshot_sha256 requires no peripheral drivers that are not already
declared by the meta-capsule's codegen, but it requires HMAC HWIP interrupts
enabled.