[ci] Temporarily remove QEMU CI
Migrating to OpenTitan earlgrey_1.0.0 broke the ability to test on upstream
QEMU, which is based on a specific OpenTitan commit before even
earlgrey_es_sival was forked.

This PR temporarily removes the QEMU CI job until the ability to run with the
lowRISC QEMU fork is debugged. The tests that were previously run with the QEMU
CI are also planned to be made available as an FPGA target in the OpenTitan
repository as part of the hardware CI effort.